Placed Items near Gridstone + Items only appear within radar
- Placed 5 items around Gridstone and resized them to a suitable size
- Location editor is now set to Gridstone
- Items now only appear within radar (15m away from player)
- Removed auto spawn mechanism and timed destroy item
