+++ FinderKeeper/Assets/Prefabs/VisibilityController 1.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d80e668de025f8e4e849e7d1e12f3b64
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
