--- FinderKeeper/Assets/Scripts/PhysControllerFP.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 4cd67567d8e221e44b8a90db17e7689d
-timeCreated: 1525632046
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
