--- FinderKeeper/Assets/Scripts/Spawn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c2e446932c7876c41bc2308f78ef994b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
