--- FinderKeeper/Assets/TimeScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd6ee77a273f8344687493bd4465c439
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
